requirements/dev.txt
pre-commit==3.4.0
pyyaml==6.0.1
virtualenv==20.24.5
black==23.9.1
pycodestyle==2.11.0
pip==23.2.1
pep8-naming==0.13.3
python-dotenv==1.0.0
django-debug-toolbar==4.2.0
